requirements.txt
--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cu129/
numpy
paddleocr[all]
paddlepaddle-gpu
pymupdf
tqdm
